Refactor keepalivego main
